+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - New http serving code in ``mpservice.http`` designed for multiple worker processes
   managed by ``uvicorn`` (as opposed to multiple worker processes managed by ``mpservice.mpserver``).
+- Removed dependency on ``asgiref``.
 
 
 ## [0.14.0] - 2023-08-22