+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## [0.14.1] - in progress
+## [0.14.1] - 2023-09-04
 
 - New http serving code in ``mpservice.http`` designed for multiple worker processes
   managed by ``uvicorn`` (as opposed to multiple worker processes managed by ``mpservice.mpserver``).